Allow dependency build to add/resolve repo
fixes https://github.com/helm/helm/issues/7214
The main issue was due to dependency build not adding the
repo causing the command to fail. The ideal response should be
to add the repo and then continue with the command,
something that dependency update does.

The functionality has been exported and then shared between
dependency build and update as a method.

Signed-off-by: Vibhav Bobade <vibhav.bobde@gmail.com>