Fix issue with sorting pod lists
The sorting previously used the selfref which contained the name
on the end for all cases except pods. In that case the selfref
was to pods and not the specific pod. This caused an issue where
multiple pods had the same selfref used as the key for softing.

The objects being sorted are tables that each have one row. In the
new setup the key is the first cells value from the first and only
row. This is the name of the resource.

Note, the Get function now requests a table. The tests have been
updated to return a Table type for the objects.

Closes #7924

Signed-off-by: Matt Farina <matt@mattfarina.com>