Add checking of length of resourceList before creating of deleting
A chart being installed which only contains CRDs and not
any templates tries to install the resources by default.
The resourceList which is used in this case does not check
if there are resources present in it or not. This commit
adds checks to those particular places where we need to check
if the size of resourceList > 0 during installation and deletion.

Signed-off-by: Vibhav Bobade <vibhav.bobde@gmail.com>
(cherry picked from commit 47abe66fd29162a71627ee4c9d26a9c3712e24e1)