Updating sprig and semver to newer versions
Note, there is an issue with a dependency of sprig changing
behavior. A test has been added with a description to catch if a
behavior breaking change of mergo is used.

See https://github.com/imdario/mergo/issues/139 for the mergo
issue and sprig for further details on handling this in the
future.

Closes  #7533

Signed-off-by: Matt Farina <matt@mattfarina.com>
Signed-off-by: Matheus Hunsche <matheus.hunsche@ifood.com.br>