Use go-yaml v3 directly
This code removes dependency on sigs.k8s.io/yaml which itself is a
mirror of github.com/ghodss/yaml.

All but one usages of ghodss/yaml were no-ops. That is they passed directly
to go-yaml v2 (gopkg.in/yaml.v2). Upgradring them to use
gopkg.in/yaml.v3 was trivial.

The one non-trivial use of ghodss/yaml was of the YAMLToJSON function. I
have copied a minimal tree of dependency functions to chartutil/yaml to
suppose just that function hosted with helm.

Upgrading to yaml.v3 supports retaining comments on round-trip
serialisation which is blocking the ability for helm package to support
the --set flag see https://github.com/helm/helm/issues/3141.

This commit pulls in the code and is pure refactoring (the next commit
fixes some minor differences in the test cases between go-yaml v2 and
v3), so the functionality from ghodss/yaml should be unchanged

Signed-off-by: Silas Davis <silas@monax.io>