fix(dep): add logic to make use of the enabled field for dependencies
addresses #7908

But as per my comment in that issue, https://github.com/helm/helm/issues/7908#issuecomment-629275316
it's possible/likely that this field wasn't intended for use.
And the implementation of unmarshaling twice comes off as a bit hacky to me. So maybe instead of this PR, it's better to just remove
the enabled field from the documentation.

-------
SIDE NOTE:
I'm not sure how to regenerate frobnitz_with_bom.tgz for the test to pass.

Signed-off-by: Jeff Knurek <knurek.stuff@gmail.com>