feat: Encode chart information into metadata
As a helm user we want to retrieve information about state of
our helm deployments, so that we can have a cluster-level dashboard for it.

Our approach was employing https://github.com/sstarcher/helm-exporter to
expose helm metadata via prometheus. Unfortunately with the chosen secret
storage backend, read access is required to get the information about deployed
chart and application versions.
Kubernetes does not allow to differentiate between different types of secrets
on a namespace, so providing read access to secrets could leak other secret
information (API Tokens, Certificate Keys, Credentials) through such an
exporter.
This change exposes chart and application versions in the metadata, so we
don't have to provide read access to secrets for helm-exporter anymore, but
helm-exporter could simply use a "list" RBAC permission to read the
metadata.

Signed-off-by: Manuel Rüger <manuel@rueg.eu>