Adding init flad to skip adding repos
The stable charts repository is nearing the end of its life. Once
that has happened those who attempt to run helm init will get an
error. This flag provides a means to still use Helm v2 after the
stable charts repository is gone.

A secondary situation is that Helm v2 can be used in environmnets
without a network connection.

Closes #7868

Signed-off-by: Matt Farina <matt@mattfarina.com>