feat(comp): Disable file completion when not valid
With Cobra 1.0, it is now possible to control when file completion
should or should not be done.  For example:
  helm list <TAB>
should not trigger file completion since 'helm list' does not accept
any arguments.

This commit disables file completion when appropriate and adds tests to
verify that file completion is properly disabled.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>