Fixing linting of templates on Windows
When the engine stored templates in the map the keys were generated
based on path and not filepath. filepath was being used in the linter
when retrieving content from the keys. On Windows the keys ended up
being different.

This change is to use path joins to create the lookup key. Since the
name path was used in the code it needed to be changed in order to
import the package.

Tests already exist and were failing on windows. This got in because
CI is not run on Windows.

Closes #6418

Signed-off-by: Matt Farina <matt@mattfarina.com>
(cherry picked from commit 11f658e2234f5d3c02212e30ba38cd579d134aaa)