feat(helm): add support for a `pre-ready` hook
Adds a new `pre-ready` hook type that runs after resources have been
created or updated, but before they're necessarily in a ready state.

This allows for charts that provision a resources using a subchart,
use a hook to configure those resources (for example, migrating a
database), and require that subchart resources be configured before
the pods in their chart are in a ready state. Using the `--wait`
flag with Helm will currently fail with these charts, making them
incompatible with the chart-testing tool.

Fixes #7740, fixes helm/chart-testing#202

Signed-off-by: Noah Lackstein <noah@lackstein.com>