Log config permission warning to default logger writer (stderr)
This PR changes the writer of the warning
```WARNING: Kubernetes configuration file is [...]```
to the writer of the default logger, which is `stderr`.

This is needed to always show the warning to the user instead of it
being parsed by other tools.

Im running `source <(helm completion zsh)` in my `.zshrc` and noticed this error on every start of a
new terminal:
```/proc/self/fd/11:1: command not found: WARNING:```
In this case it is not even visible to the user that helm caused the issue.
With this PR the full warning is displayed to the user instead.

We could also use the logger `log.Warn`? in `checkPerms` directly.

Signed-off-by: Luca Berneking <luca@berneking.net>