fix(helm): Don't discard pre-releases upon index merge
Previously, when attempting to merge a new version into an existing
index.yaml which includes overlapping release and pre-release versions
(such as 0.1.0 and 0.1-dev), the merge would silently discard the
pre-release version from the newly merged index.

Fix this by performing an exact semantic comparison of the versions from
the provided index file.

Signed-off-by: Joe Stringer <joe@cilium.io>