fix: rollback to release containing resource with resource-policy=keep annotation
When rollback to a previous release revision that contains a resource
having the annotaion 'helm.sh/resource-policy=keep', while the current
release revision dosen't have this resource, rollback will fail with
an error message:
'Error: no <Kind> with the name "<Resource name>" found'

First of all the error message is misleading, since the resource
actually exists. But the resource is not part of the current manifest
(it is left there in a previous update due to the resource-policy=keep
annotation). So helm cannot create a patch to update the given resource
from the current to the target manifest, and fails with an error.

The logic is updated to in this scenario (resource is part of the target
manifest but NOT part of the current manifest) skip the upgrade IF the
resource-policy=keep annotation is set.

The error message is also updated and the the logic for checking the
resource-policy annotation is broken out into a separate function.

fix #8228

Signed-off-by: Ted Petersson <ted.petersson@ericsson.com>