Fix action/List.Run to skip the latest release filter if ListSuperseded is set
If superseded releases should be returned, then do not filter for only the
latest releases, because the latest releases exclude superseded releases, by
definition.

Previously, the logic was: if _only_ superseded releases should be included,
then do not filter for only the latest releases. That is incorrect; for
example, when _all_ releases were requested, only the latest releases would be
returned.

Signed-off-by: Daniel Lipovetsky <dlipovetsky@d2iq.com>