fix(pkg/chartutil) Remove disabled dependencies
Remove aliased dependencies that are disabled with a condition.
Before, if a chart had multiple aliased dependencies that reference the
same chart, the metadata was shared between all dependencies. This led
to dependencies not being processed correctly and resulted in a bug
where disabled dependencies were not removed and therefore rendered.

The metadatas dependencies are now deep-copied to prevent metadata
corruption.

Signed-off-by: Olivier Michaelis <38879457+oliviermichaelis@users.noreply.github.com>