Allow for helm install script to work on Windows
This installs to the GOPATH bin folder which will work cross platform and will likely be setup on the user's path. Previously it used a standard unix path which is non-existent on Windows. As currently written it requires the user to have go installed. To run on windows you then just need to pass the --no-sudo flag.

Signed-off-by: mburst <mburst@maxburstein.com>