feat(helm): add --reset-then-reuse-values flag to 'helm upgrade'
When '--reset-then-reuse-values' is used on 'helm upgrade', the chart's values will be
reset to the values of the deployed chart while the current release's values will be
reused and merged with the values passed as argument (is any). '--reset-values' and
'--reuse-values' flags take precedence over `--reset-then-reuse-values', making it
ignored if one or the other is also used.

Closes #8085, #3957

Signed-off-by: Quentin Devos <quentin@devos.pm>