Remove verbose logging when updating chart repo index
This commit removes logs lines resulting from loading the index of a
chart repo containing many charts that don't adhere to semver.

Motivation behind this change is simple, there is little to no impact
from these log lines, and in most cases no action that the user could
reasonably expect to make.  In the authors case, this change reduces
the output of `helm repo up` from +100MB to <1KB.

Signed-off-by: Nathan Schmidt <nathan.schmidt@imc.com>