fix(template) Do not error when both template show-only and output-dir
Previously, when one set --show-only and --output-dir together, helm template
would throw an error and all files were rendered.

Refactor some of the code so that manifests and hooks are filtered in a central
location and only render manifests and hooks that match --show-only.

Additionally, use StringSliceVarP for the --show-only flag.

Fixes #9722

Signed-off-by: Andreas Karis <ak.karis@gmail.com>