Adds urlutil.CanonicalAddr and use it when comparing chart repository and chart url
If a chart repository was added with https://myhost/chart and the
corresponding index.yaml file points to
https://myhost:443/helm/somechart-1.0.0.tgz the newly introduced
security check in v3.6.1 will mark the two hostname and scheme
comparison as different.
Canonicalize the hostname by always adding the port, if not defined, based
upon the used scheme.
http  maps to '80'
https maps to '443'

Signed-off-by: Lüchinger Dominic <dev@snowgarden.ch>