Improve handling of the credential forward check
Improve SchemeHostAndPortMatches in urlutil to ensure that credentials
will be used / forwarded if chart url and repo url are the same.
The same now means, that an additional check has been added which ensures,
that the implicit ports of http and https (80/443) will be used in the
compare check if they are not given, so that
http(s)://foo.bar:443/ == http(s)://foo.bar/. Old behavior
detected a diff here and then people needed to add a --pass-credentials
flag, which is a really dangerous option if you don't know exactly
what you are doing.

Signed-off-by: Felix Becker <git@felixbecker.name>