feat(plugin): Allow plugins to receive empty arg
Empty arguments were being swallowed by the plugin parsing logic; this
commit fixes that.

This was not a big problem as empty arguments are probably quite rare.
However, the shell completion logic does have a use case for an empty
argument.  Plugins using Cobra have a '__complete' command that can be
called to obtain the valid completion choices.  That command takes as
arguments what the user has typed.  When the user has not typed anything
the command needs to be passed an explicit empty argument.  For example:

  helm 2to3 __complete ''

but since the empty argument is not being passed to the plugin we get a
missing argument error because the __complete command requires at least
one argument.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>