fix: reinstall previously uninstalled chart with --keep-history
Before, if a chart was uninstalled using the `--keep-history` flag, the
next attempt to upgrade/install the chart would result in an error.

This commit fixes that by correctly checking if the last release of the
chart stored in history has been uninstalled and running the install
with the `--replace` flag if that's the case.

Signed-off-by: Alex Petrov <alex.petrov.vt@gmail.com>