Add --hook-parallelism flag
Add --hook-parallelism flag to helm deployment hooks but not the testing hook

What this PR does / why we need it:
Adds hook parallelism with a default of 1 as a flag for applicable commands. Does not add parallelism for test hooks as that seems to be the reason why this https://github.com/helm/community/pull/165 was closed but I think there is still value having deployment hooks work in parallel.

Special notes for your reviewer:
Based off of https://github.com/helm/helm/pull/8946 by abaehremc and
https://github.com/helm/helm/pull/7792 by akhilles