[chart/loader] use strict yaml unmarshaling for chart files
This updates the helm chart loader to use strict yaml unmarshaling (UnmarshalStrict) instead of the non-strict mode (Unmarshal). The main benefit of this is to prevent duplicate keys within the Values.yaml file which seems like that would never be intentionally done by the user. Right now if there are duplicate keys within Values.yaml, the last one will be used and the other ones will be silently ignored.

Signed-off-by: Aadesh Patel <aadeshdpatel@gmail.com>