Use persistent `RESTClientGetter` in CLI
This changes the `RESTClientGetter` to one which caches the underlying
clients after first initialization.

This reduces memory usage for all commands, with observed reductions
from ~4MiB for `helm get all`, to ~10MiB for `helm install` (using
https://github.com/stefanprodan/podinfo).

For charts which make use of Custom Resource Definitions, this is
likely even higher as the client will not be initialized twice (one
time for the installation process of the CRDs, the second time for the
actual installation of the chart).

Follow up on #11886

Signed-off-by: Hidde Beydals <hidde@hhh.computer>