Updating the Helm maintainers
These maintainers have made tremendous contributions but are
currently inactive. I have connected with each of them and they
are focused on other work, at present. They all confirmed moving
to an emeritus status. If they ever want to return to active
development/maintainer-ship of Helm, I would be delighted.

The governance has a process for inactive folks and, as there are
items that need to be voted on, inactive maintainer make decision
making difficult.

These maintainers are and will continue to be missed.

Signed-off-by: Matt Farina <matt.farina@suse.com>