improve get-helm-3 GH api interaction
This commit solves two problems:  The Github API has extremely low
rate-limits on requests that could cause issues if you have a large
CI/CD farm behind a NAT gateway or similar that causes a large amount of
requests to appear to come from the same IP.  By utilizing the
environment variable GITHUB_TOKEN if present, CI/CD can pass in a GH
personal access token to bypass severe rate limiting.

Additionally, the Github API is not determinisitic on the return format
of its JSON -- it may or may not be compressed and/or single line.  We
should try to more accurately extract the tag name value instead of
assuming multi-line output.

Signed-off-by: Brandon Ewing <brandon.ewing@warningg.com>